api/requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.38.0
python-dotenv==1.2.1
httpx==0.28.1
pydantic>=2.12.5
pydantic-settings>=2.6.0
motor>=3.6.0

# File handling
python-multipart>=0.0.20
python-docx>=1.1.0
PyPDF2>=3.0.0

# Authentication
bcrypt>=4.0.0
python-jose[cryptography]>=3.4.0
email-validator>=2.0.0

# Production: Circuit Breaker & Resilience
pybreaker>=1.0.0

# Testing
requests>=2.31.0

# Production: Monitoring & Observability
prometheus-client>=0.20.0  # Metrics export
python-json-logger>=2.0.0  # Structured logging
